+++ src/components/table/table.vue
@@ -21,7 +21,7 @@
     <div :class="getTableClasses.getTableHeight">
       <table aria-describedby="describe" class="spr-w-full spr-table-fixed" cellspacing="0" cellpadding="0">
         <thead>
-          <tr>
+          <tr v-if="props.removeHeaderOnEmpty && sortedData.length > 0">
             <th v-for="(header, keyHeader) in headers" :key="keyHeader" :class="[getTableClasses.headerClasses]">
               <div :class="getTableClasses.headerNameClass">
                 <span :class="[{ 'spr-cursor-pointer': header.sort }]" @click="header.sort && sortData(header.field)">
@@ -64,8 +64,27 @@
                   :variant="column.avatarVariant ? column.avatarVariant : 'initial'"
                   :initial="sortedData[keyIndex][column.field].title"
                 />
+                <div
+                  v-if="column.hasIcon"
+                  class="spr-flex spr-items-center spr-rounded-full spr-bg-mushroom-200 spr-p-1"
+                >
+                  <Icon :icon="sortedData[keyIndex][column.field].icon || ''" />
+                </div>
                 <div>
+                  <div v-if="column.hasLozengeTitle" class="spr-mt-1">
+                    <spr-lozenge
+                      :label="sortedData[keyIndex][column.field].title"
+                      :tone="sortedData[keyIndex][column.field].lozengeTone || 'plain'"
+                      :url="sortedData[keyIndex][column.field].lozengeAvatarUrl"
+                      :fill="sortedData[keyIndex][column.field].lozengeFill || false"
+                    >
+                      <template v-if="sortedData[keyIndex][column.field].lozengeIcon" #icon>
+                        <Icon :icon="sortedData[keyIndex][column.field].lozengeIcon || ''" />
+                      </template>
+                    </spr-lozenge>
+                  </div>
                   <div
+                    v-else
                     :class="[
                       'spr-text-color-strong spr-font-size-200 spr-font-normal',
                       { 'spr-text-color-strong spr-body-sm-regular-medium': column.hasSubtext },
@@ -117,6 +136,7 @@ import SprAvatar from '@/components/avatar/avatar.vue';
 import SprEmptyState from '@/components/empty-state/empty-state.vue';
 import SprBadge from '@/components/badge/badge.vue';
 import SprTableActions from '@/components/table/table-actions/table-actions.vue';
+import SprLozenge from '@/components/lozenge/lozenge.vue';
 
 import { tablePropTypes, tableEmitTypes } from './table';
 import { useTable } from './use-table';