+++ docs/guide/changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.13.1 (2025-03-19)
+- Feat
+  - Added Table Lozenge Cells ([#b2ea23e](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/b2ea23e8af842870b5bc8b3e97918a825394f7fd?refName=refs/heads/2025/FEATURE/Xyrk/TABLE_ROW_ICON_LOZENGE_AND_HEADER) by @jfabula)
+
+## 1.13.0 (2025-03-17)
+
+- Feat
+  - Add collapsible component and documentation. Update card props and documentation ([#b5d669ad](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/commit/b5d669adb78ab6a41351130270bfa3621b5309b5?refName=refs%2Fheads%2F2025%2Ffeature%2Fjames%2Fcollapsible) by @jmanalang)
+
 ## 1.12.5 (2025-03-17)
 
 - Fix